Replace script templates with worker module
