+++ AppTaskQueue/appscale/taskqueue/tq_config.py
@@ -52,11 +52,8 @@ queue:
   TEMPLATE_DIR = os.path.join(
     os.path.dirname(sys.modules['appscale.taskqueue'].__file__), 'templates')
 
-  # The location of a header of a queue worker script.
-  HEADER_LOC = os.path.join(TEMPLATE_DIR, 'header.py')
-  
-  # The location of the task template code.
-  TASK_LOC = os.path.join(TEMPLATE_DIR, 'task.py')
+  # The worker script for Celery to use.
+  WORKER_MODULE = 'appscale.taskqueue.push_worker'
 
   # XML configs use "-" while yaml uses "_". These are the tags which 
   # need to be converted to match the yaml tags.
@@ -238,41 +235,6 @@ queue:
 
     logger.debug("XML queue info is {0}".format(converted))
     return converted
-
-  def create_celery_worker_scripts(self):
-    """ Creates the task worker python script. It uses a configuration file
-    for setup.
-
-    Returns:
-      The full path of the worker script.
-    """
-    header_template = file_io.read(self.HEADER_LOC)
-    task_template = file_io.read(self.TASK_LOC)
-    header_template = header_template.replace("APP_ID", self._app_id)
-    script = header_template.replace("CELERY_CONFIGURATION", self._app_id) + \
-      '\n'
-
-    for name, queue in self.queues.iteritems():
-      # Celery only handles push queues.
-      if not isinstance(queue, PushQueue):
-        continue
-
-      # The queue name is used as a function name so replace invalid chars
-      queue_name = queue.name.replace('-', '_')
-      new_task = task_template.\
-        replace("QUEUE_NAME", self.get_queue_function_name(queue_name))
-      # For tasks generated by mapreduce, or destined to be run by a module,
-      # the hostname may have a prefix that corresponds to a different
-      # subdomain.
-      # AppScale does not support that type of routing currently, so the main
-      # loadbalancer IP/hostname is used here for the execution of a task.
-      new_task = new_task.\
-        replace("PUBLIC_IP", "\"{}\"".format(self.get_public_ip()))
-      script += new_task + '\n'
-
-    worker_file = self.get_celery_worker_script_path(self._app_id)
-    file_io.write(worker_file, script)
-    return worker_file
   
   @staticmethod
   def remove_config_files(app_id):
@@ -281,9 +243,7 @@ queue:
     Args:
       app_id: The application ID.
     """
-    worker_file = TaskQueueConfig.get_celery_worker_script_path(app_id)
     config_file = TaskQueueConfig.get_celery_configuration_path(app_id)
-    file_io.delete(worker_file)
     file_io.delete(config_file)
 
   @staticmethod
