+++ AppTaskQueue/setup.py
@@ -12,7 +12,7 @@ setup(
   platforms='Posix',
   install_requires=[
     'cassandra-driver',
-    'celery<4.0.0',
+    'celery>=3.1,<4.0.0',
     'PyYaml',
     'tornado==4.2.0'
   ],
@@ -30,6 +30,5 @@ setup(
     'console_scripts': [
       'appscale-taskqueue=appscale.taskqueue.appscale_taskqueue:main'
     ]
-  },
-  package_data={'appscale.taskqueue': ['templates/*']}
+  }
 )
